# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Museum Ticketing & Booking Platform API
  version: 0.2.0
  description: Bilingual (EN/AM) online booking, payment, and gate check-in for the
    Zoological Natural History Museum's digital ticketing track.
paths:
  /api/v1/admin/audit-log/:
    get:
      operationId: listAuditLog
      description: |-
        `GET /admin/audit-log` -- Platform Admin only, read-only
        (FR-AUDIT-001). The append-only trail every module writes through
        `core.services.write_audit_log`, filterable and paginated.
      parameters:
      - in: query
        name: action
        schema:
          type: string
        description: Filter by action, e.g. `booking.cancelled`.
      - in: query
        name: actor
        schema:
          type: string
          format: uuid
        description: Filter by acting account UUID.
      - in: query
        name: entityId
        schema:
          type: string
        description: Filter by the audited entity's id.
      - in: query
        name: entityType
        schema:
          type: string
        description: Filter by the audited entity's type, e.g. `account`.
      - in: query
        name: from
        schema:
          type: string
          format: date-time
        description: Only entries created at/after this ISO-8601 instant.
      - name: limit
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - name: offset
        required: false
        in: query
        description: The initial index from which to return the results.
        schema:
          type: integer
      - in: query
        name: to
        schema:
          type: string
          format: date-time
        description: Only entries created at/before this ISO-8601 instant.
      tags:
      - v1
      security:
      - TokenVersionAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedAuditLogEntryList'
          description: ''
  /api/v1/admin/staff/:
    get:
      operationId: v1_admin_staff_list
      description: |-
        `GET /admin/staff` and `POST /admin/staff` -- both Platform Admin
        only (FR-ACC-002).
      parameters:
      - name: limit
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - name: offset
        required: false
        in: query
        description: The initial index from which to return the results.
        schema:
          type: integer
      tags:
      - v1
      security:
      - TokenVersionAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedAccountList'
          description: ''
    post:
      operationId: v1_admin_staff_create
      description: |-
        `GET /admin/staff` and `POST /admin/staff` -- both Platform Admin
        only (FR-ACC-002).
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StaffCreate'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/StaffCreate'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/StaffCreate'
        required: true
      security:
      - TokenVersionAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
          description: ''
  /api/v1/admin/staff/{id}/:
    put:
      operationId: v1_admin_staff_update
      description: |-
        `PUT /admin/staff/{id}` and `DELETE /admin/staff/{id}` -- both
        Platform Admin only (FR-ACC-002). No `GET` here: Document 04 has no
        single-staff-account retrieve endpoint, only the list at
        `/admin/staff` (mirrors `apps.catalog.CategoryDetailView`).
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StaffUpdate'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/StaffUpdate'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/StaffUpdate'
      security:
      - TokenVersionAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
          description: ''
    delete:
      operationId: v1_admin_staff_destroy
      description: |-
        `PUT /admin/staff/{id}` and `DELETE /admin/staff/{id}` -- both
        Platform Admin only (FR-ACC-002). No `GET` here: Document 04 has no
        single-staff-account retrieve endpoint, only the list at
        `/admin/staff` (mirrors `apps.catalog.CategoryDetailView`).
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - v1
      security:
      - TokenVersionAuth: []
      responses:
        '204':
          description: No response body
...
components:
  schemas:
    Account:
      type: object
      description: |-
        `UserProfile` (Document 04) -- Visitor and Staff share one shape; a
        Visitor's password-adjacent fields are simply never populated.

        `email_verified_at`/`phone_verified_at` are exposed (read-only) so the
        frontend can gate a Visitor's flow -- e.g. requiring email verification
        right after OTP confirmation (FR-ACC-003) -- without having to guess at
        verification state or discover it only when a later action (like
        `create_booking`) rejects it.
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        email:
          type: string
          readOnly: true
        phone:
          type: string
          readOnly: true
          nullable: true
        full_name:
          type: string
          readOnly: true
        role:
          allOf:
          - $ref: '#/components/schemas/AccountRoleEnum'
          readOnly: true
        language_preference:
          allOf:
          - $ref: '#/components/schemas/LanguagePreferenceEnum'
          readOnly: true
        active:
          type: boolean
          readOnly: true
        email_verified_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        phone_verified_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        created_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - active
      - created_at
      - email
      - email_verified_at
      - full_name
      - id
      - language_preference
      - phone
      - phone_verified_at
      - role
    AccountRoleEnum:
      enum:
      - visitor
      - cashier
      - museum_manager
      - platform_admin
      type: string
      description: |-
        * `visitor` - Visitor
        * `cashier` - Cashier
        * `museum_manager` - Museum Manager
        * `platform_admin` - Platform Admin
...
    AuditLogEntry:
      type: object
      description: |-
        `AuditLogEntry` (Document 04) -- one append-only row of the platform
        audit trail (FR-AUDIT-001). Read-only: entries are written by
        `core.services.write_audit_log` throughout the codebase, never through
        the API.
      properties:
        id:
          type: integer
          readOnly: true
        actorUserId:
          type: string
          format: uuid
          readOnly: true
          nullable: true
        action:
          type: string
          readOnly: true
        entityType:
          type: string
          readOnly: true
        entityId:
          type: string
          readOnly: true
        metadata:
          readOnly: true
        createdAt:
          type: string
          format: date-time
          readOnly: true
      required:
      - action
      - actorUserId
      - createdAt
      - entityId
      - entityType
      - id
      - metadata
...
    LanguagePreferenceEnum:
      enum:
      - en
      - am
      type: string
      description: |-
        * `en` - English
        * `am` - Amharic
...
    PaginatedAccountList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Account'
        meta:
          type: object
          properties:
            limit:
              type: integer
            offset:
              type: integer
            total:
              type: integer
          required:
          - limit
          - offset
          - total
      required:
      - data
      - meta
    PaginatedAuditLogEntryList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/AuditLogEntry'
        meta:
          type: object
          properties:
            limit:
              type: integer
            offset:
              type: integer
            total:
              type: integer
          required:
          - limit
          - offset
          - total
      required:
      - data
      - meta
...
    StaffCreate:
      type: object
      description: |-
        `StaffCreateRequest` (Document 04) -- FR-ACC-002, Platform Admin
        only. No `password` field: services.create_staff_account provisions
        the account with no usable password and sends a set-password link
        instead (FR-ACC-006's flow, reused).
      properties:
        email:
          type: string
          format: email
        phone:
          type: string
          default: ''
          maxLength: 32
        full_name:
          type: string
          maxLength: 255
        role:
          $ref: '#/components/schemas/StaffRoleEnum'
      required:
      - email
      - full_name
      - role
...
    StaffRoleEnum:
      enum:
      - cashier
      - museum_manager
      type: string
      description: |-
        * `cashier` - Cashier
        * `museum_manager` - Museum Manager
    StaffUpdate:
      type: object
      description: |-
        `StaffUpdateRequest` (Document 04) -- FR-ACC-002, Platform Admin
        only. Every field optional so a single `PUT` can update the staff
        profile and role without changing the password.
      properties:
        email:
          type: string
          format: email
        phone:
          type: string
          nullable: true
          maxLength: 32
        full_name:
          type: string
          maxLength: 255
        role:
          $ref: '#/components/schemas/StaffRoleEnum'
        active:
          type: boolean
...
  securitySchemes:
    TokenVersionAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT